Extract popover footer
